moved all the internal streams that are needed externally to util
